fix: update donation count
