+++ apps/web/src/api/dashboard.ts
@@ -97,23 +97,7 @@ export function useDashboardSummary(storeId?: string) {
   });
 }
 
-export interface CharityImpact {
-  openingBalance: number;
-  totalRaised: number;
-  totalDonated: number;
-  pendingPayout: number;
-  bookingContributions: number;
-  annualCap: number;
-  annualDonated: number;
-}
-
-export function useCharityImpact() {
-  return useQuery<CharityImpact>({
-    queryKey: ['dashboard', 'charity-impact'],
-    queryFn: () => api.get<CharityImpact>('/dashboard/charity-impact'),
-    staleTime: 5 * 60_000,
-  });
-}
+export { useCharityImpact, type CharityImpact } from './impact.js';
 
 export interface CharityDonationRow {
   id: string;
